+++ test/callback.test.ts
@@ -1,21 +1,19 @@
 import assert from "assert"
-import weak from "../src/weak"
+import { addCallback, callbacks, create, removeCallback, removeCallbacks } from "../src/weak"
 import { gc } from "expose-ts-gc"
 
-describe('weak()', function () {
-
+describe("create()", function () {
   afterEach(gc)
 
-  describe('garbage collection callback', function () {
-
-    it('should accept a function as second argument', function () {
-      var r = weak({}, function () {})
-      assert.equal(1, weak.callbacks(r).length)
+  describe("garbage collection callback", function () {
+    it("should accept a function as second argument", function () {
+      const r = create({}, function () {})
+      assert.equal(1, callbacks(r).length)
     })
 
-    it('should invoke the callback before the target is gc\'d', function () {
-      var called = false
-      weak({}, function () {
+    it("should invoke the callback before the target is gc'd", function () {
+      let called = false
+      create({}, function () {
         called = true
       })
       assert(!called)
@@ -23,15 +21,14 @@ describe('weak()', function () {
       assert(called)
     })
 
-    it('should invoke *all* callbacks in the internal "callback" Array'
-    , function () {
-      var r = weak({})
-        , called1 = false
-        , called2 = false
-      weak.addCallback(r, function () {
+    it('should invoke *all* callbacks in the internal "callback" Array', function () {
+      let r = create({}),
+        called1 = false,
+        called2 = false
+      addCallback(r, function () {
         called1 = true
       })
-      weak.addCallback(r, function () {
+      addCallback(r, function () {
         called2 = true
       })
       gc()
@@ -39,63 +36,61 @@ describe('weak()', function () {
       assert(called2)
     })
 
-    it('should preempt code for GC callback but not nextTick callbacks'
-    , function(done) {
-      var calledGcCallback = false
-      , calledTickCallback = false
-      weak({}, function() {
+    it("should preempt code for GC callback but not nextTick callbacks", function (done) {
+      let calledGcCallback = false,
+        calledTickCallback = false
+      create({}, function () {
         calledGcCallback = true
       })
 
-      process.nextTick(function() {
+      process.nextTick(function () {
         calledTickCallback = true
-      });
+      })
 
       assert(!calledGcCallback)
       assert(!calledTickCallback)
       gc()
       assert(calledGcCallback)
       assert(!calledTickCallback)
-      setTimeout(function() {
-        assert(calledTickCallback);
-        done();
+      setTimeout(function () {
+        assert(calledTickCallback)
+        done()
       }, 0)
     })
-
   })
 })
 
-describe('callbacks()', function () {
-
+describe("callbacks()", function () {
   it('should return the Weakref\'s "callback" Array', function () {
-    var r = weak({}, function() {})
-      , callbacks = weak.callbacks(r)
-    assert(Array.isArray(callbacks))
-    assert.equal(1, callbacks.length)
+    const r = create({}, function () {}),
+      cbs = callbacks(r)
+    assert(Array.isArray(cbs))
+    assert.equal(1, cbs.length)
   })
-
 })
 
-describe('removeCallback()', function() {
-
-  it('removed callbacks should not be called', function() {
-    var called = false
-      , fn = function() { called = true }
-      , r = weak({}, fn)
-    weak.removeCallback(r, fn)
+describe("removeCallback()", function () {
+  it("removed callbacks should not be called", function () {
+    let called = false,
+      fn = function () {
+        called = true
+      },
+      r = create({}, fn)
+    removeCallback(r, fn)
     gc()
     assert(!called)
   })
-
 })
 
-describe('removeCallbacks()', function() {
-
-  it('removed callbacks should not be called', function() {
-     var called = false, fn = function() { called = true }, r = weak({}, fn)
-     weak.removeCallbacks(r)
-     gc()
-     assert(!called)
+describe("removeCallbacks()", function () {
+  it("removed callbacks should not be called", function () {
+    let called = false,
+      fn = function () {
+        called = true
+      },
+      r = create({}, fn)
+    removeCallbacks(r)
+    gc()
+    assert(!called)
   })
-
 })