+++ test/promise.test.ts
@@ -1,14 +1,14 @@
 import assert from "assert"
-import weak from "../src/weak"
+import { isWeakRef } from "../src/weak"
 
-describe('promise', function () {
+describe("promise", function () {
   if (!global.Promise) {
     // Version of JS without `Promise`; nothing to do.
-    return;
+    return
   }
 
-  it('should not mistake a promise for a weak reference', function () {
-    var p = new Promise(function () {});
-    assert(!weak.isWeakRef(p));
-  });
+  it("should not mistake a promise for a weak reference", function () {
+    const p = new Promise(function () {})
+    assert(!isWeakRef(p))
+  })
 })